refactor: sort components
