+++ client/src/components/molecules/MoviePreview/MoviePreview.js
