+++ client/src/components/molecules/MoviePreview/__tests__/MoviePreview.spec.js
