+++ client/src/components/molecules/MoviePreview/index.js
